Latest API changes

+++ src/main/java/com/ecom/imgur/controller/ImageController.java
@@ -1,9 +1,13 @@
 package com.ecom.imgur.controller;
 
+import com.ecom.imgur.common.Constant;
+import com.ecom.imgur.exception.UserException;
+import com.ecom.imgur.exception.UserNotFoundException;
 import com.ecom.imgur.model.UserAccountResponse;
 import com.ecom.imgur.service.ImageServiceImpl;
 import com.ecom.imgur.model.ImageResponse;
 import com.ecom.imgur.model.ImagesResponse;
+import com.ecom.imgur.service.UserServiceImpl;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
@@ -19,6 +23,8 @@ public class ImageController {
     @Autowired
     private ImageServiceImpl imageService;
 
+    @Autowired
+    private UserServiceImpl userService;
     /**
      *
      * @param username
@@ -26,21 +32,32 @@ public class ImageController {
      */
 
     @GetMapping("/view/{username}")
-    public ResponseEntity<ImagesResponse> viewImage(@PathVariable String username) {
-        ImagesResponse images=imageService.getImages(username);
-        log.info("Successfully list of Images sent "+images.getImages());
-        return ResponseEntity.ok().body(images);
+    public ResponseEntity<ImagesResponse> viewImage(@PathVariable String username,@RequestHeader String userID,@RequestHeader String password) {
+
+        if(userService.authenticateUser(userID ,password))
+        {
+            ImagesResponse images=imageService.getImages(username);
+            log.info("Successfully list of Images sent "+images.getImages());
+            return ResponseEntity.ok().body(images);
+        }else {
+            throw new UserException(Constant.USER_AUTH_FAILED);
+        }
+
     }
     /**
      *
      * @return Uploaded image ImageResponse object
      */
     @PostMapping("/upload/")
-    public ResponseEntity<ImageResponse> uploadImage( @RequestParam("file") MultipartFile file) {
+    public ResponseEntity<ImageResponse> uploadImage( @RequestParam("file") MultipartFile file,@RequestHeader String userID,@RequestHeader String password) {
         log.info("Upload image file name {} ",file.getName());
-        ImageResponse imgurApiResponse= imageService.saveImage(file);
-        log.info("Successfully image Upload  {} for ",file.getName());
-        return ResponseEntity.ok(imgurApiResponse);
+        if(userService.authenticateUser(userID ,password)) {
+            ImageResponse imgurApiResponse = imageService.saveImage(file);
+            log.info("Successfully image Upload  {} for ", file.getName());
+            return ResponseEntity.ok(imgurApiResponse);
+        }else {
+            throw new UserException(Constant.USER_AUTH_FAILED);
+        }
     }
 
     /**
@@ -49,10 +66,14 @@ public class ImageController {
      * @return deleted imageId
      */
     @DeleteMapping("/delete/{imageId}")
-    public ResponseEntity<String> deleteImageForAuthenticatedUser(@PathVariable String imageId) {
+    public ResponseEntity<String> deleteImageForAuthenticatedUser(@PathVariable String imageId,@RequestHeader String userID,@RequestHeader String password) {
+        if(userService.authenticateUser(userID ,password)) {
         imageService.deleteImage(imageId);
         log.info("{} Image Successfully deleted", imageId);
         return ResponseEntity.ok(imageId+ "Image Successfully deleted");
+        }else {
+            throw new UserException(Constant.USER_AUTH_FAILED);
+        }
     }
     /**
      *
